same width inbox card and contact card
