+++ app/contact/page.tsx
@@ -1,36 +1,8 @@
-'use client';
-import React, { useState } from 'react';
-import { HeaderDiv, Main, MainDiv, Tab, TabDiv, Title, TopDiv } from './style';
-import PersonList from './personList';
-import CompanyList from './companyList';
+import React from 'react';
+import Contact from './contact';
 
-export default function Contact() {
-  const [activeTab, setActiveTab] = useState('People');
-  const tabItem = ['People', 'Company'];
-  return (
-    <Main>
-      <MainDiv>
-        <TopDiv>
-          <HeaderDiv>
-            <Title>Contact</Title>
-            <TabDiv>
-              {tabItem.map((tab) => (
-                <Tab
-                  key={tab}
-                  active={activeTab === tab}
-                  onClick={() => setActiveTab(tab)}
-                >
-                  {tab}
-                </Tab>
-              ))}
-            </TabDiv>
-          </HeaderDiv>
-        </TopDiv>
-        <div>
-          {activeTab === 'People' && <PersonList />}
-          {activeTab === 'Company' && <CompanyList />}
-        </div>
-      </MainDiv>
-    </Main>
-  );
+function ContactPage() {
+  return <Contact />;
 }
+
+export default ContactPage;
