+++ app/details/[ticket_id]/ticketDetails.tsx
@@ -306,143 +306,145 @@ function TicketDetails(props: Props) {
             </ButtonDiv>
           </StatusDiv>
         </TopDiv>
-        <BottomDiv>
-          <CenterDiv>
-            <ActivityDiv>
-              <Avatar
-                imgSrc={
-                  'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
-                }
-                name={''}
-                size={20}
-              />
-              <QuestionCard
-                title={'@Aniket can you please look into this?'}
-                time={'3 day ago'}
-              />
-            </ActivityDiv>
-            <LineDiv />
-            <ActivityDiv>
-              <Avatar
-                imgSrc={
-                  'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
-                }
-                name={''}
-                size={20}
-              />
-              <MessageCard
-                title={'Sanjay send email'}
-                time={'2 days ago'}
-                subTitle={'To Teamcamp Support '}
-                message={
-                  "Hey,Thank you for choosing our services through our partner, Parthern. To ensure you receive the full benefits of your purchase, we invite you to create an account with us at Teamcamp.Create Your Account Today!Setting up your Teamcamp account is quick and easy. Follow this link to get started: www.teamcamp.app Need Help?If you have any questions or need assistance during the registration process, please do not hesitate to reply to this emailWe're excited to have you on board and look forward to supporting your project management needs!Warm regards,Sanjay M."
-                }
-              />
-            </ActivityDiv>
-            <LineDiv />
-            <ActivityDiv>
-              <Avatar
-                imgSrc={
-                  'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
-                }
-                name={''}
-                size={20}
-              />
-              <Message>
-                Connect AI <span>set priority to</span> Low
-                <SVGIcon
-                  name='dot-icon'
-                  width='4'
-                  height='4'
-                  fill='none'
-                  viewBox='0 0 4 4'
-                />
-                <span>2 min ago</span>
-              </Message>
-            </ActivityDiv>
-            <LineDiv />
-            <ActivityDiv>
-              <Avatar
-                imgSrc={
-                  'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
-                }
-                name={''}
-                size={20}
-              />
-              <Message>
-                Connect AI <span>assigned this ticket to</span> Sanjay M.
-                <SVGIcon
-                  name='dot-icon'
-                  width='4'
-                  height='4'
-                  fill='none'
-                  viewBox='0 0 4 4'
-                />
-                <span>2 min ago</span>
-              </Message>
-            </ActivityDiv>
-          </CenterDiv>
-          <InputDiv>
-            <Avatar
-              imgSrc={
-                'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
-              }
-              size={20}
-              name={''}
-            />
-            <Input modeSelectedItem={modeSelectedItem}>
-              <RichTextBox />
-              <InputIcon>
-                <DropDownWithTag
-                  onClick={handleMessageModeTag}
-                  selectedValue={modeSelectedItem}
-                  dropdownOpen={messageModeDropdown}
-                  title='Email'
-                  onClose={() => {
-                    setMessageModeDropdown(false);
-                  }}
-                  items={modeItem}
-                  onChange={handleModeItemChange}
-                  isTag={true}
-                  iconName={modeSelectedItem?.icon}
-                  iconSize='12'
-                  iconViewBox='0 0 12 12'
-                  isActive={true}
-                  className={
-                    submenuPosition === 'upwards'
-                      ? 'submenu-upwards'
-                      : 'submenu-downwards'
+        <div style={{ padding: '0 20px' }}>
+          <BottomDiv>
+            <CenterDiv>
+              <ActivityDiv>
+                <Avatar
+                  imgSrc={
+                    'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
                   }
-                  onMouseEnter={(e: any) =>
-                    handleMouseEnter(e, setSubmenuPosition)
-                  }
-                  dropDownStyle={{ maxWidth: 142, width: '100%' }}
-                  tagStyle={{
-                    backgroundColor: 'unset',
-                  }}
+                  name={''}
+                  size={20}
                 />
-                <IconDiv modeSelectedItem={modeSelectedItem}>
-                  <Icon
-                    onClick={() => {}}
-                    iconName='attach-icon'
-                    iconSize='12'
-                    iconViewBox='0 0 12 12'
-                    size={true}
+                <QuestionCard
+                  title={'@Aniket can you please look into this?'}
+                  time={'3 day ago'}
+                />
+              </ActivityDiv>
+              <LineDiv />
+              <ActivityDiv>
+                <Avatar
+                  imgSrc={
+                    'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
+                  }
+                  name={''}
+                  size={20}
+                />
+                <MessageCard
+                  title={'Sanjay send email'}
+                  time={'2 days ago'}
+                  subTitle={'To Teamcamp Support '}
+                  message={
+                    "Hey,Thank you for choosing our services through our partner, Parthern. To ensure you receive the full benefits of your purchase, we invite you to create an account with us at Teamcamp.Create Your Account Today!Setting up your Teamcamp account is quick and easy. Follow this link to get started: www.teamcamp.app Need Help?If you have any questions or need assistance during the registration process, please do not hesitate to reply to this emailWe're excited to have you on board and look forward to supporting your project management needs!Warm regards,Sanjay M."
+                  }
+                />
+              </ActivityDiv>
+              <LineDiv />
+              <ActivityDiv>
+                <Avatar
+                  imgSrc={
+                    'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
+                  }
+                  name={''}
+                  size={20}
+                />
+                <Message>
+                  Connect AI <span>set priority to</span> Low
+                  <SVGIcon
+                    name='dot-icon'
+                    width='4'
+                    height='4'
+                    fill='none'
+                    viewBox='0 0 4 4'
                   />
-                  <Icon
-                    onClick={() => {}}
-                    iconName='send-icon'
+                  <span>2 min ago</span>
+                </Message>
+              </ActivityDiv>
+              <LineDiv />
+              <ActivityDiv>
+                <Avatar
+                  imgSrc={
+                    'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
+                  }
+                  name={''}
+                  size={20}
+                />
+                <Message>
+                  Connect AI <span>assigned this ticket to</span> Sanjay M.
+                  <SVGIcon
+                    name='dot-icon'
+                    width='4'
+                    height='4'
+                    fill='none'
+                    viewBox='0 0 4 4'
+                  />
+                  <span>2 min ago</span>
+                </Message>
+              </ActivityDiv>
+            </CenterDiv>
+            <InputDiv>
+              <Avatar
+                imgSrc={
+                  'https://firebasestorage.googleapis.com/v0/b/teamcamp-app.appspot.com/o/UserProfiles%2FUntitled1_1701236653470.jpg?alt=media&token=8bc07cdb-5fcc-4c69-8e0d-c9978b94b3e4'
+                }
+                size={20}
+                name={''}
+              />
+              <Input modeSelectedItem={modeSelectedItem}>
+                <RichTextBox />
+                <InputIcon>
+                  <DropDownWithTag
+                    onClick={handleMessageModeTag}
+                    selectedValue={modeSelectedItem}
+                    dropdownOpen={messageModeDropdown}
+                    title='Email'
+                    onClose={() => {
+                      setMessageModeDropdown(false);
+                    }}
+                    items={modeItem}
+                    onChange={handleModeItemChange}
+                    isTag={true}
+                    iconName={modeSelectedItem?.icon}
                     iconSize='12'
                     iconViewBox='0 0 12 12'
-                    size={true}
                     isActive={true}
-                    className='icon'
+                    className={
+                      submenuPosition === 'upwards'
+                        ? 'submenu-upwards'
+                        : 'submenu-downwards'
+                    }
+                    onMouseEnter={(e: any) =>
+                      handleMouseEnter(e, setSubmenuPosition)
+                    }
+                    dropDownStyle={{ maxWidth: 142, width: '100%' }}
+                    tagStyle={{
+                      backgroundColor: 'unset',
+                    }}
                   />
-                </IconDiv>
-              </InputIcon>
-            </Input>
-          </InputDiv>
-        </BottomDiv>
+                  <IconDiv modeSelectedItem={modeSelectedItem}>
+                    <Icon
+                      onClick={() => {}}
+                      iconName='attach-icon'
+                      iconSize='12'
+                      iconViewBox='0 0 12 12'
+                      size={true}
+                    />
+                    <Icon
+                      onClick={() => {}}
+                      iconName='send-icon'
+                      iconSize='12'
+                      iconViewBox='0 0 12 12'
+                      size={true}
+                      isActive={true}
+                      className='icon'
+                    />
+                  </IconDiv>
+                </InputIcon>
+              </Input>
+            </InputDiv>
+          </BottomDiv>
+        </div>
       </MainDiv>
       <ProfileSection />
     </Main>
