+++ app/details/style.tsx
@@ -29,7 +29,6 @@ const BottomDiv = styled.div`
   display: flex;
   flex-direction: column;
   height: calc(100vh - 86px);
-  padding: 0 20px;
 `;
 
 const MainDiv = styled.div`
